+++ packages/icon/src/js/_build/_icons.js
@@ -900,6 +900,10 @@ export default [
     category: "social",
   },
   {
+    name: "social-x-twitter-circle",
+    category: "social",
+  },
+  {
     name: "social-you-tube-circle",
     category: "social",
   },