+++ packages/private-design-output/gulpfile.js
@@ -1,36 +1,4 @@
-const { src, dest, parallel } = require("gulp");
-const sass = require("gulp-dart-sass");
-const babel = require("gulp-babel");
-const uglify = require("gulp-uglify");
-const rename = require("gulp-rename");
-const rollup = require("@rollup/stream");
-const { nodeResolve } = require("@rollup/plugin-node-resolve");
-const cjs = require("@rollup/plugin-commonjs");
-const source = require("vinyl-source-stream");
-const buffer = require("vinyl-buffer");
-
-// SCSS include paths
-const scssPaths = ["node_modules", "node_modules/@uqds"];
-
-// Build and export SCSS
-function compileSCSS() {
-  return src("./src/scss/**/*.scss")
-    .pipe(
-      sass({
-        includePaths: scssPaths,
-        outputStyle: "compressed",
-      }).on("error", sass.logError)
-    )
-    .pipe(dest("./dist/css"));
-}
-
-// Export resources
-/*
-function exportFontFiles() {
-  return src([])
-  .pipe(dest('./dist/fonts'));
-}
-*/
+const { src, dest, parallel } = require("gulp"); // eslint-disable-line
 
 function exportImages() {
   return src(["./src/images/*.*"]).pipe(dest("./dist/images"));
@@ -46,52 +14,24 @@ function exportHtml() {
 }
 
 // Temporary slidemenu addition
-function exportSlidmenu() {
+function exportSlidemenu() {
   return src("./src/js/slide-menu.js").pipe(dest("./dist/js"));
 }
 
-// Bundle Javascript modules
-/**
- * TODO: Babel presets/targets need setting if clients have browser support
- * issues
- */
-function bundleJS() {
-  return rollup({
-    input: "./src/js/main.js",
-    output: {
-      format: "iife",
-      name: "uq",
-    },
-    plugins: [nodeResolve(), cjs()],
-  })
-    .pipe(source("uqds.js"))
-    .pipe(buffer())
-    .pipe(
-      babel({
-        presets: ["@babel/preset-env"],
-        plugins: ["@babel/plugin-proposal-class-properties"],
-      })
-    )
-    .pipe(dest("./dist/js"))
-    .pipe(rename("uqds.min.js"))
-    .pipe(uglify())
-    .pipe(dest("./dist/js"));
-}
-
-exports.compileSCSS = compileSCSS;
-//exports.exportFontFiles = exportFontFiles;
 exports.exportImages = exportImages;
 exports.exportFavicon = exportFavicon;
 exports.exportHtml = exportHtml;
-// Temporary slidemenu addition
-exports.exportSlidmenu = exportSlidmenu;
-exports.bundleJS = bundleJS;
-// Temporary slidemenu addition
+exports.exportSlidmenu = exportSlidemenu;
 exports.default = parallel(
-  compileSCSS,
   exportImages,
   exportFavicon,
   exportHtml,
-  exportSlidmenu,
-  bundleJS
+  exportSlidemenu
+);
+
+exports.prepare = parallel(
+  exportImages,
+  exportFavicon,
+  exportHtml,
+  exportSlidemenu
 );