+++ packages/storybook-html/stories/components/footer-legacy/footer.stories.js
@@ -4,18 +4,12 @@ import "./footer.scss";
 // import HTML template strings
 import footerHTML from "./footer.html";
 
-import docs from "./footer.docs.mdx";
-
 export default {
-  title: "Legacy/Footer",
+  title: "Deprecated/Footer",
   parameters: {
     layout: "fullscreen",
     backgrounds: {
       default: "UQ Neutral 1",
-    },
-    docs: {
-      page: docs,
-      inlineStories: false,
     },
     previewTabs: {
       canvas: { hidden: false },