+++ packages/storybook-html/stories/components/section/section.stories.js
@@ -1,18 +1,8 @@
 import classNames from "classnames";
-import docs from "./section.docs.mdx";
-import "../../../../sections/src/scss/main.scss";
+import "@uqds/sections/src/scss/main.scss";
 
 export default {
   title: "Components/Section",
-  parameters: {
-    docs: {
-      page: docs,
-    },
-    layout: "fullscreen",
-    previewTabs: {
-      canvas: { hidden: false },
-    },
-  },
   argTypes: {
     title: {
       control: "text",
@@ -51,13 +41,13 @@ export const Section = {
 };
 
 export const Shaded = {
+  ...Section,
   args: {
     title: "Title",
     description: "Description",
     content: "",
     shaded: true,
   },
-  render: Section.render,
 };
 
 export const Container = {
@@ -88,12 +78,11 @@ export const Container = {
 };
 
 export const ShadedContainer = {
-  name: "Shaded with container",
+  ...Container,
   args: {
     title: "Title",
     description: "Description",
     content: "",
     shaded: true,
   },
-  render: Container.render,
 };