+++ packages/storybook-html/stories/components/site-header/site-header.stories.js
@@ -10,15 +10,9 @@ import siteHeaderNav from "@uqds/site-header/src/js/site-header-navigation";
 import siteHeaderWithSubnavHTML from "./site-header-with-subnav.html";
 import siteHeaderWithoutSubnavHTML from "./site-header-without-subnav.html";
 
-import docs from "./site-header.docs.mdx";
-
 export default {
   title: "Components/Site header",
   parameters: {
-    docs: {
-      page: docs,
-      inlineStories: false,
-    },
     layout: "fullscreen",
     previewTabs: {
       canvas: { hidden: false },
@@ -26,22 +20,26 @@ export default {
   },
 };
 
-export const siteHeaderWithSubnav = () => {
-  useEffect(() => {
-    var navElem = document.getElementById("jsNav");
-    new siteHeaderNav(navElem, "uq-site-header__navigation");
-  });
-  return siteHeaderWithSubnavHTML;
+export const siteHeaderWithSubnav = {
+  render: () => {
+    useEffect(() => {
+      var navElem = document.getElementById("jsNav");
+      new siteHeaderNav(navElem, "uq-site-header__navigation");
+    });
+    return siteHeaderWithSubnavHTML;
+  },
+
+  name: "With sub-navigation",
 };
 
-siteHeaderWithSubnav.storyName = "With sub-navigation";
+export const siteHeaderWithoutSubnav = {
+  render: () => {
+    useEffect(() => {
+      var navElem = document.getElementById("jsNav");
+      new siteHeaderNav(navElem, "uq-site-header__navigation");
+    });
+    return siteHeaderWithoutSubnavHTML;
+  },
 
-export const siteHeaderWithoutSubnav = () => {
-  useEffect(() => {
-    var navElem = document.getElementById("jsNav");
-    new siteHeaderNav(navElem, "uq-site-header__navigation");
-  });
-  return siteHeaderWithoutSubnavHTML;
+  name: "Without sub-navigation",
 };
-
-siteHeaderWithoutSubnav.storyName = "Without sub-navigation";