+++ packages/storybook-html/stories/layout/container/container.stories.js
@@ -1,17 +1,10 @@
-import docs from "./container.docs.mdx";
+import "@uqds/layout/src/scss/main.scss";
 import "../layout-demo.scss";
-import "../../../../layout/src/scss/main.scss";
 
 export default {
   title: "Layout/Container",
   parameters: {
     layout: "fullscreen",
-    docs: {
-      page: docs,
-    },
-    previewTabs: {
-      canvas: { hidden: false },
-    },
   },
 };