+++ packages/storybook-html/stories/layout/page-layouts/sidebar/sidebar.stories.js
@@ -1,17 +1,10 @@
-import docs from "./sidebar.docs.mdx";
+import "@uqds/layout/src/scss/main.scss";
 import "../../layout-demo.scss";
-import "../../../../../layout/src/scss/main.scss";
 
 export default {
   title: "Layout/Page layouts/Sidebar",
   parameters: {
     layout: "fullscreen",
-    docs: {
-      page: docs,
-    },
-    previewTabs: {
-      canvas: { hidden: false },
-    },
   },
 };