+++ packages/storybook-html/stories/templates/level2.stories.js
@@ -4,6 +4,8 @@ import { footer } from "../components/footer/footer.stories";
 import { Breadcrumb } from "../components/breadcrumb/breadcrumb.stories";
 import { Level2 as Hero } from "../components/hero/hero.stories";
 import { LocalNav } from "../components/navigation/local-nav/local-nav.stories";
+import "@uqds/layout/src/scss/main.scss";
+import "../layout/layout-demo.scss";
 
 export default {
   title: "Templates/Level 2",
@@ -28,7 +30,7 @@ export default {
 export const Level2 = {
   render: ({ showGrid }) => `
 <div class="${classNames({ "uqds-layout-demo-bg": showGrid })}">
-  ${Header()}
+  ${Header.render()}
   ${Breadcrumb.render()}
   <div class="uq-sections">
     ${Hero.render(Hero.args)}
@@ -84,6 +86,6 @@ export const Level2 = {
       </div>
     </div>
   </div>
-  ${footer()}
+  ${footer.render()}
 </div>`,
 };